pyproject.toml
[tool.mypy]
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
python_version = "3.9"
strict_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Ash Arnold <ash@dragonpaw.org>"]
description = "Dragonpaw's Discord Bot"
license = "Apache"
name = "dragonpaw-bot"
version = "2.0.0"

[tool.poetry.dependencies]
aiohttp = "^3.8.1"
emojis = "^0.6.0"
hikari = {extras = ["speedups"], version = "^2.0.0-alpha.105"}
hikari-lightbulb = "^2.1.3"
palettable = "^3.3.0"
pydantic = "^1.9.0"
python = "3.9.4"
python-dotenv = "^0.15.0"
safer = "^4.4.0"
toml = "^0.10.2"
uvloop = "^0.16.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipython = "^8.0.1"
isort = "^5.7.0"
mypy = "^0.931"
pylint = "^2.6.0"
types-PyYAML = "^6.0.3"
types-toml = "^0.10.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
